Admin Pages (Content)
A few administrative views to modify content with.  How it is now, the
user will be prompted to select the section they wish to modify, then be
taken to a form where they can alter the text and save it.

Please have at least one entry in the database for content before trying
to test it, else you'll probably get an error.